refactor: rename `uiohook` to `hook` and move devices hooks to `hook/`
